Organize transform-related functions into src/profile-logic/transforms.js
